+++ pysys-examples/internal/testcases/PySys_internal_096/run.py
@@ -1,22 +1,13 @@
+import os, sys
 import pysys
 from pysys.constants import *
 from pysys.basetest import BaseTest
-import os, sys
 
 class PySysTest(BaseTest):
 
 	def execute(self):
-		self.pythonDocTest(os.path.dirname(pysys.__file__)+'/utils/stringutils.py', 
+		self.pythonDocTest(os.path.dirname(pysys.__file__)+'/process/user.py', 
 			pythonPath=sys.path)
 			
 	def validate(self):
-		return
-		# ensure these appear at start of the line, which for some CI writers is important
-		self.assertGrep('pysys.out', expr='^stdoutPrint-CUSTOMWRITER-setup')
-		self.assertGrep('pysys.out', expr='^stdoutPrint-CUSTOMWRITER-processResult')
-		self.assertGrep('pysys.out', expr='^stdoutPrint-CUSTOMWRITER-setup')
-		
-		self.assertGrep('pysys.out', expr='^sys.stdout.write-CUSTOMWRITER-setup')
-		self.assertGrep('pysys.out', expr='^sys.stdout.write-CUSTOMWRITER-processResult')
-		self.assertGrep('pysys.out', expr='^sys.stdout.write-CUSTOMWRITER-setup')
-		
+		pass # all validation is done by pythonDocTest